+++ docs/changes.md
@@ -3,7 +3,7 @@
 All notable changes to this project will be documented in this file.
 
 
-## Unreleased (YYYY-MM-DD)
+## 1.0.5 (2023-04-12)
 
 - Remove ``autoadd_bindir.cancel``-function